requirements.txt
# Core dependencies - minimal versions that work across Python 3.9-3.11
fastapi>=0.109.1
uvicorn>=0.35.0
pydantic>=2.11.9
starlette>=0.27.0

# ML & AI
torch>=2.0.0
transformers>=4.30.0
huggingface-hub>=0.20.0

# Vector DB
qdrant-client>=1.1.1

# Testing
pytest>=7.0.0
pytest-asyncio>=0.20.0
httpx>=0.24.0

# Code Quality
black>=23.0.0
flake8>=6.0.0

# Utils
python-multipart>=0.0.5
typing-extensions>=4.0.0

# Install python-dotenv
python-dotenv>=1.0.0